requirements.txt
chainmap==1.0.3
click==7.1.2
combomethod==1.0.12
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nulltype==2.3.1
options==1.4.10
PyMySQL==0.10.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.12.0
Werkzeug==1.0.1
